refactor: improve PDF output, prep for Codespaces
This commit overhauls the project template to optimize for PDF output,
improve the initial project setup, and begin to add support for
GitHub Codespaces.

Key changes include a custom 'answer-box' environment with a Lua filter
and a tuned `_quarto.yml` for high-quality PDFs with the xelatex
engine. The initialization script is now more robust by finding files
dynamically, and the README setup is simplified using a template file.
A new `docs/` folder organizes assignment materials, and a
`.devcontainer` is included for a reproducible Docker environment.